vkd3d: Use vkd3d_atomic_increment_u32() in d3d12_pipeline_state_AddRef().

+++ libs/vkd3d/vkd3d_private.h
@@ -1288,7 +1288,7 @@ struct d3d12_pipeline_uav_counter_state
 struct d3d12_pipeline_state
 {
     ID3D12PipelineState ID3D12PipelineState_iface;
-    LONG refcount;
+    unsigned int refcount;
 
     union
     {
